training/action-recognition/requirements.txt
# 核心依賴
# YOLOv11 需要較新的 Ultralytics 版本，鎖定到已知穩定的大版本範圍
ultralytics>=8.2.0,<9.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 資料處理
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
PyYAML>=6.0

# 視覺化和監控
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0

# 工具和實用程式
tqdm>=4.64.0
psutil>=5.9.0
pandas>=1.3.0

# 可選：Jupyter 支援
jupyter>=1.0.0
ipywidgets>=7.6.0

# 可選：資料增強
albumentations>=1.3.0

# 可選：模型優化
onnx>=1.12.0
onnxruntime>=1.12.0
